requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-genai
praw
pandas
openpyxl
